Header Style
The body has 80% of the width size. THe header has a bck color set and the content of the header has a 15% margin in the left and right. The header contents are organized with flex property, with space between. The menu li in the right is organized as flex-row, and has a specific gap of 20px between the menu links.