The call to action design
The call to action div has a 70% width, and a blue background. The content inside it has a 60% width and it is flex row with space between. The h3 and h4 are align left and have no space between. The button on the right, is positioned in center and has a bcg color and other properties set as above.